gitlabl: Fix warnings that are causing the gitlab breaks
1. Fix the formatting string used to print the command for md.c
2. Remove non-standard use of strcasecmp (not in the C spec) so this
   was causing a failure in the GitLab environment.

BRANCH=none
BUG=none
TEST=twister -T zephyr/test/drivers/

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3869002
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>

+++ zephyr/test/drivers/default/src/console_cmd/md.c
@@ -36,7 +36,8 @@ ZTEST_USER(console_cmd_md, test_default_count)
 	uint8_t memory[] = { 0x01, 0x02, 0x03, 0x04 };
 	char cmd[128] = { 0 };
 
-	zassume_true(sprintf(cmd, "md %llu", memory) != 0, NULL);
+	zassume_true(sprintf(cmd, "md %" PRIuPTR, (uintptr_t)memory) != 0,
+		     NULL);
 	zassert_ok(shell_execute_cmd(get_ec_shell(), cmd), NULL);
 }
 
@@ -45,7 +46,8 @@ ZTEST_USER(console_cmd_md, test_count_arg)
 	uint8_t memory[] = { 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08 };
 	char cmd[128] = { 0 };
 
-	zassume_true(sprintf(cmd, "md %llu 2", memory) != 0, NULL);
+	zassume_true(sprintf(cmd, "md %" PRIuPTR " 2", (uintptr_t)memory) != 0,
+		     NULL);
 	zassert_ok(shell_execute_cmd(get_ec_shell(), cmd), NULL);
 }
 
@@ -54,7 +56,8 @@ ZTEST_USER(console_cmd_md, test_byte_format)
 	uint8_t memory[] = { 0x01, 0x02, 0x03, 0x04 };
 	char cmd[128] = { 0 };
 
-	zassume_true(sprintf(cmd, "md .b %llu", memory) != 0, NULL);
+	zassume_true(sprintf(cmd, "md .b %" PRIuPTR, (uintptr_t)memory) != 0,
+		     NULL);
 	zassert_ok(shell_execute_cmd(get_ec_shell(), cmd), NULL);
 }
 
@@ -63,7 +66,8 @@ ZTEST_USER(console_cmd_md, test_half_format)
 	uint8_t memory[] = { 0x01, 0x02, 0x03, 0x04 };
 	char cmd[128] = { 0 };
 
-	zassume_true(sprintf(cmd, "md .h %llu", memory) != 0, NULL);
+	zassume_true(sprintf(cmd, "md .h %" PRIuPTR, (uintptr_t)memory) != 0,
+		     NULL);
 	zassert_ok(shell_execute_cmd(get_ec_shell(), cmd), NULL);
 }
 
@@ -72,6 +76,8 @@ ZTEST_USER(console_cmd_md, test_string_format)
 	char memory[] = "hello world";
 	char cmd[128] = { 0 };
 
-	zassume_true(sprintf(cmd, "md .s %llu 12", memory) != 0, NULL);
+	zassume_true(sprintf(cmd, "md .s %" PRIuPTR " 12", (uintptr_t)memory) !=
+			     0,
+		     NULL);
 	zassert_ok(shell_execute_cmd(get_ec_shell(), cmd), NULL);
 }
